+++ game_patch/main/main.cpp
@@ -43,6 +43,7 @@
 
 GameConfig g_game_config;
 HMODULE g_hmodule;
+std::mt19937 g_rng;
 
 CallHook<void()> rf_init_hook{
     0x004B27CD,
@@ -269,6 +270,13 @@ void init_logging()
     xlog::info("Command line: {}", GetCommandLineA());
 }
 
+void initialize_random_generator()
+{
+    // seed rng with the current time
+    auto seed = std::chrono::steady_clock::now().time_since_epoch().count();
+    g_rng.seed(static_cast<unsigned long>(seed));
+}
+
 void log_system_info()
 {
     try {
@@ -337,6 +345,9 @@ extern "C" DWORD __declspec(dllexport) Init([[maybe_unused]] void* unused)
     // Init logging and crash dump support first
     init_logging();
     init_crash_handler();
+
+    // Init random number generator
+    initialize_random_generator();
 
     // Enable Data Execution Prevention
     if (!SetProcessDEPPolicy(PROCESS_DEP_ENABLE))