+++ game_patch/multi/server_internal.h
@@ -33,6 +33,7 @@ struct ServerAdditionalConfig
     VoteConfig vote_extend;
     VoteConfig vote_restart;
     VoteConfig vote_next;
+    VoteConfig vote_rand;
     VoteConfig vote_previous;
     int spawn_protection_duration_ms = 1500;
     std::optional<float> spawn_life;
@@ -50,6 +51,7 @@ struct ServerAdditionalConfig
     std::optional<float> max_fov;
     int anticheat_level = 0;
     bool stats_message_enabled = true;
+    bool random_rotation = false;
     std::string welcome_message;
     bool weapon_items_give_full_ammo = false;
     float kill_reward_health = 0.0f;
@@ -69,6 +71,7 @@ void init_server_commands();
 void extend_round_time(int minutes);
 void restart_current_level();
 void load_next_level();
+void load_rand_level();
 void load_prev_level();
 void server_vote_on_limbo_state_enter();
 void process_delayed_kicks();